requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
absl-py==1.4.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==2.1.1
cycler==0.11.0
dadaptation==3.2
dominate==2.7.0
ema-pytorch==0.4.5
fonttools==4.38.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grad-cam==1.5.0
grpcio==1.51.3
idna==3.4
imageio==2.22.4
importlib-metadata==6.0.0
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
lightning-utilities==0.9.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
monai==1.0.1
networkx==2.6.3
nibabel==4.0.2
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image
scikit-learn
scipy==1.7.3
seaborn==0.13.1
SimpleITK==2.2.1
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchmetrics==1.0.2
torchsummary==1.5.1
torchsummaryX==1.3.0
torch-ema
torchvision==0.11.2+cu111
tornado==6.2
tqdm==4.64.1
ttach==0.0.3
typing_extensions==4.4.0
unzip==1.0.0
urllib3==1.26.12
visdom==0.2.3
websocket-client==1.4.2
Werkzeug==2.2.3
xgboost
zipp==3.15.0
